Commiting to merge. A few name changes and drop handler separated.
